Added timed waves and audio
Enemies now spawn in timed intervals. This can be set in the inspector of BigBadSingleton in the GameplayManager script component.

Game audio was added (combat SFX) and background music responds to if the player is in or out of combat.

If player dies, the game freezes. There is an event handler here were a game over UI canvas can be called for the player to restart or quit the game.