Added new enemy
New enemy type (bomber) has been added to the EnemyPalette. This enemy moves toward the player quickly and explodes with a large radius of damage (enables a CircleCollider2D component which if the player is in when the enemy explodes, will receive damage).

Changed the enemy wave spawn logic to spawn certain enemies depending on which wavenumber it is.