Completed Boss Fight Mechanics
Boss has both melee and ranged (projectile) attacks

Boss changes to an 'enraged' state when at 50% health. This changes the projectile sprite and firing type. This also changes the animation states of the boss to look more angry.

Trigger collider has been placed at the start of the boss room (just as the player enters) which triggers the boss fight (stops spawning normal enemies and plays the combat music).